refactor(bpmn-model): rename task validator class
