+++ bpmn-model/src/main/java/io/camunda/zeebe/model/bpmn/validation/zeebe/BusinessRuleTaskValidator.java
@@ -16,47 +16,47 @@
 package io.camunda.zeebe.model.bpmn.validation.zeebe;
 
 import io.camunda.zeebe.model.bpmn.impl.ZeebeConstants;
+import io.camunda.zeebe.model.bpmn.instance.BusinessRuleTask;
 import io.camunda.zeebe.model.bpmn.instance.ExtensionElements;
-import io.camunda.zeebe.model.bpmn.instance.ScriptTask;
-import io.camunda.zeebe.model.bpmn.instance.zeebe.ZeebeScript;
+import io.camunda.zeebe.model.bpmn.instance.zeebe.ZeebeCalledDecision;
 import io.camunda.zeebe.model.bpmn.instance.zeebe.ZeebeTaskDefinition;
 import java.util.Collection;
 import org.camunda.bpm.model.xml.validation.ModelElementValidator;
 import org.camunda.bpm.model.xml.validation.ValidationResultCollector;
 
-public final class ScriptTaskValidation implements ModelElementValidator<ScriptTask> {
+public final class BusinessRuleTaskValidator implements ModelElementValidator<BusinessRuleTask> {
 
   @Override
-  public Class<ScriptTask> getElementType() {
-    return ScriptTask.class;
+  public Class<BusinessRuleTask> getElementType() {
+    return BusinessRuleTask.class;
   }
 
   @Override
   public void validate(
-      final ScriptTask element, final ValidationResultCollector validationResultCollector) {
+      final BusinessRuleTask element, final ValidationResultCollector validationResultCollector) {
 
     if (!hasExactlyOneExtension(element)) {
       validationResultCollector.addError(
           0,
           String.format(
               "Must have either one 'zeebe:%s' or one 'zeebe:%s' extension element",
-              ZeebeConstants.ELEMENT_SCRIPT, ZeebeConstants.ELEMENT_TASK_DEFINITION));
+              ZeebeConstants.ELEMENT_CALLED_DECISION, ZeebeConstants.ELEMENT_TASK_DEFINITION));
     }
   }
 
-  private boolean hasExactlyOneExtension(final ScriptTask element) {
+  private boolean hasExactlyOneExtension(final BusinessRuleTask element) {
     final ExtensionElements extensionElements = element.getExtensionElements();
 
     if (extensionElements == null) {
       return false;
     }
 
-    final Collection<ZeebeScript> scriptExtensions =
-        extensionElements.getChildElementsByType(ZeebeScript.class);
+    final Collection<ZeebeCalledDecision> calledDecisionExtensions =
+        extensionElements.getChildElementsByType(ZeebeCalledDecision.class);
     final Collection<ZeebeTaskDefinition> taskDefinitionExtensions =
         extensionElements.getChildElementsByType(ZeebeTaskDefinition.class);
 
-    return scriptExtensions.size() == 1 && taskDefinitionExtensions.isEmpty()
-        || scriptExtensions.isEmpty() && taskDefinitionExtensions.size() == 1;
+    return calledDecisionExtensions.size() == 1 && taskDefinitionExtensions.isEmpty()
+        || calledDecisionExtensions.isEmpty() && taskDefinitionExtensions.size() == 1;
   }
 }
