+++ bpmn-model/src/test/java/io/camunda/zeebe/model/bpmn/validation/BusinessRuleTaskValidatorTest.java
@@ -26,7 +26,7 @@ import io.camunda.zeebe.model.bpmn.instance.zeebe.ZeebeTaskDefinition;
 import java.util.function.Consumer;
 import org.junit.jupiter.api.Test;
 
-class BusinessRuleTaskValidationTest {
+class BusinessRuleTaskValidatorTest {
 
   @Test
   void emptyDecisionId() {
