+++ engine/src/test/java/io/camunda/zeebe/engine/processing/deployment/model/validation/BusinessRuleTaskValidatorTest.java
@@ -14,7 +14,7 @@ import io.camunda.zeebe.model.bpmn.instance.zeebe.ZeebeCalledDecision;
 import java.util.function.Consumer;
 import org.junit.jupiter.api.Test;
 
-public class BusinessRuleTaskValidationTest {
+public class BusinessRuleTaskValidatorTest {
 
   @Test
   void emptyDecisionIdExpression() {
